requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.2
django-allauth==65.1.0
django-environ==0.11.2
django-requests==0.0.2
idna==3.10
jwt==1.3.1
pillow==11.0.0
pycparser==2.22
pytz==2024.2
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
urllib3==2.2.3
